+++ src/hooks.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
 namespace hooksH {
+    extern int selectedStartpos;
+    extern std::vector<StartPosObject*> startPositions;
     void switchStartPos(int incBy, bool direction = true);
 }